fix bug api

+++ src/main/java/fpt/custome/yourlure/dto/dtoOut/AdminDiscountVoucherDetailDtoOut.java
@@ -19,8 +19,8 @@ public class AdminDiscountVoucherDetailDtoOut {
 //    private Integer used;
     private Date start_date;
     private Date end_date;
-    private Integer discountValue;
-    private Integer minSpentAmount;
+    private Float discountValue;
+    private Float minSpentAmount;
     private Integer minCheckoutItemsQuantity;
 //    private Integer applyOncePerCustomer;
